* Fixed Typo in RGB  * Fixed a type error
git-svn-id: file:///home/michael/svn/yuni/trunk@1398 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/samples/core/io/01.threadediterator/main.cpp
@@ -46,7 +46,7 @@ protected:
 		--pCounter;
 	}
 
-	virtual Flow onFile(const String&, const String&, const String& name, size_t size)
+	virtual Flow onFile(const String&, const String&, const String& name, uint64 size)
 	{
 		printSpaces();
 		std::cout << "  -  " << name << " (" << size << " bytes)" << std::endl;
